shell model
